backend/requirements.txt
fastapi
uvicorn
httpx
langchain
langchain-groq
langchain-core
python-dotenv
beautifulsoup4
pydantic